# Mark some Integer methods as inline