```
$ benchmark-driver -v --rbenv 'before;after;before --jit;after --jit' benchmark/mjit_integer.yml --repeat-count=4
before: ruby 2.8.0dev (2020-06-27T05:43:59Z master 609ac6ee2c) [x86_64-linux]
after: ruby 2.8.0dev (2020-06-27T09:40:50Z integer-builtin 72e5e698e7) [x86_64-linux]
before --jit: ruby 2.8.0dev (2020-06-27T05:43:59Z master 609ac6ee2c) +JIT [x86_64-linux]
after --jit: ruby 2.8.0dev (2020-06-27T09:40:50Z integer-builtin 72e5e698e7) +JIT [x86_64-linux]
Calculating -------------------------------------
                         before       after  before --jit  after --jit
        mjit_abs(-1)    38.801M     41.313M       52.936M      71.152M i/s -     40.000M times in 1.030890s 0.968213s 0.755626s 0.562176s
mjit_bit_length(100)    39.179M     40.442M       51.317M      69.839M i/s -     40.000M times in 1.020947s 0.989061s 0.779463s 0.572749s
       mjit_even?(2)    39.377M     41.804M       52.152M      71.191M i/s -     40.000M times in 1.015819s 0.956838s 0.766986s 0.561868s
    mjit_integer?(0)    40.345M     42.618M       53.870M      72.188M i/s -     40.000M times in 0.991454s 0.938567s 0.742522s 0.554112s
  mjit_magnitude(-1)    38.826M     41.377M       53.099M      71.561M i/s -     40.000M times in 1.030227s 0.966709s 0.753311s 0.558963s
        mjit_odd?(1)    39.588M     41.618M       53.023M      71.880M i/s -     40.000M times in 1.010412s 0.961123s 0.754383s 0.556485s
         mjit_ord(1)    40.491M     42.898M       54.263M      69.175M i/s -     40.000M times in 0.987870s 0.932450s 0.737152s 0.578244s
        mjit_to_i(1)    40.457M     42.451M       53.951M      70.862M i/s -     40.000M times in 0.988704s 0.942254s 0.741413s 0.564477s
      mjit_to_int(1)    40.131M     42.803M       52.942M      70.067M i/s -     40.000M times in 0.996730s 0.934514s 0.755542s 0.570883s
       mjit_zero?(0)    43.260M     42.977M       73.035M      71.392M i/s -     40.000M times in 0.924650s 0.930732s 0.547683s 0.560286s

Comparison:
                     mjit_abs(-1)
         after --jit:  71152147.0 i/s
        before --jit:  52936259.5 i/s - 1.34x  slower
               after:  41313204.7 i/s - 1.72x  slower
              before:  38801409.7 i/s - 1.83x  slower

             mjit_bit_length(100)
         after --jit:  69838607.6 i/s
        before --jit:  51317383.1 i/s - 1.36x  slower
               after:  40442410.4 i/s - 1.73x  slower
              before:  39179324.0 i/s - 1.78x  slower

                    mjit_even?(2)
         after --jit:  71191085.5 i/s
        before --jit:  52152198.7 i/s - 1.37x  slower
               after:  41804340.6 i/s - 1.70x  slower
              before:  39377078.2 i/s - 1.81x  slower

                 mjit_integer?(0)
         after --jit:  72187631.3 i/s
        before --jit:  53870493.3 i/s - 1.34x  slower
               after:  42618163.7 i/s - 1.69x  slower
              before:  40344780.7 i/s - 1.79x  slower

               mjit_magnitude(-1)
         after --jit:  71561052.4 i/s
        before --jit:  53098897.2 i/s - 1.35x  slower
               after:  41377484.5 i/s - 1.73x  slower
              before:  38826385.0 i/s - 1.84x  slower

                     mjit_odd?(1)
         after --jit:  71879778.1 i/s
        before --jit:  53023449.6 i/s - 1.36x  slower
               after:  41617978.5 i/s - 1.73x  slower
              before:  39587802.7 i/s - 1.82x  slower

                      mjit_ord(1)
         after --jit:  69174967.5 i/s
        before --jit:  54262866.1 i/s - 1.27x  slower
               after:  42897762.2 i/s - 1.61x  slower
              before:  40491145.4 i/s - 1.71x  slower

                     mjit_to_i(1)
         after --jit:  70862014.1 i/s
        before --jit:  53951013.1 i/s - 1.31x  slower
               after:  42451391.1 i/s - 1.67x  slower
              before:  40457017.6 i/s - 1.75x  slower

                   mjit_to_int(1)
         after --jit:  70066924.7 i/s
        before --jit:  52942135.0 i/s - 1.32x  slower
               after:  42802999.9 i/s - 1.64x  slower
              before:  40131243.5 i/s - 1.75x  slower

                    mjit_zero?(0)
        before --jit:  73034937.7 i/s
         after --jit:  71392080.1 i/s - 1.02x  slower
              before:  43259592.2 i/s - 1.69x  slower
               after:  42976929.0 i/s - 1.70x  slower
```